Rename pickups.py -> pickup.py
